requirements.txt
openai
pdfplumber
python-dotenv
azure-ai-formrecognizer
